Start building creature chromosome.

+++ Assets/Sandbox/Tile.cs
@@ -3,11 +3,12 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Tile {
+public class Tile
+{
     public float elevation { get; private set; }
-    public int temperature { get; private set; }
+    public float temperature { get; private set; }
     public float lightLevel { get; private set; }
-    public TerrainData terrain { get; private set; }   
+    public TerrainData terrain { get; private set; }
     public Vector2Int position { get; private set; }
 
     public HashSet<Actor> actors;
@@ -16,7 +17,7 @@ public class Tile {
     private Level level;
     private List<Tile> adjacentTiles;
 
-    public Tile(Level level, Vector2Int position, TerrainData terrain, float elevation, int temperature, float lightLevel)
+    public Tile(Level level, Vector2Int position, TerrainData terrain, float elevation, float temperature, float lightLevel)
     {
         this.level = level;
         this.position = position;
